+++ docs/mkdocs/blog/changelog.md
@@ -5,10 +5,13 @@ This changelog was not introduced until v0.2.12, so many of the critical release
 Version 0.2 improved greatly on 0.1, with a greater focus on the user experience, plugins, local performance, and a whole lot more. Read the release notes below for some of the highlights.
 
 ### v0.2.19
-- **Added `show columns` action**  
+- **Added `show columns` action.**  
   The action `show columns` will now display a pipe's columns and data types.
-- **Refactoring**  
+- **`docker-compose` bugfix.**  
+  When `docker-compose` is installed globally, skip using the virtual environment version.
+- **Refactoring / linting**  
   A lot of code was cleaned up to conform with cleaner programming practices.
+
 
 ### v0.2.18
 - **Added `login` action.**